requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.1.4
django-crispy-forms==1.7.2
django-formtools==2.2
django-import-export==2.5.0
django-redis==4.4.0
django-reversion==3.0.9
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
MarkupPy==1.14
numpy==1.20.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.2
pycparser==2.20
pycrypto==2.6.1
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
qrcode==6.1
redis==3.5.3
requests==2.25.1
six==1.15.0
tablib==3.0.0
urllib3==1.26.4
xlrd==2.0.1
xlwt==1.3.0